No more processors (#26)
* Replace Processor with tool.user_input_received

* Integrate no more processors with prior tables changes

* Update results so that only approved rows are shown

* Corpus update

Change in which LambdaProcessors and Processors more broadly are deprecated. Rather than gating code to be run once user input is received inside functions passed to Processors, we add a method `user_input_received` to the Tool class, and expect this code to be gated by if statements using this method as a guard.